# server/gen/swagger/swagger.yaml
basePath: /api/v1
definitions:
  apierror.APIError:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
...
  queryparam.PaginationMeta:
    properties:
      has_next:
        type: boolean
      has_prev:
        type: boolean
      limit:
        type: integer
      page:
        type: integer
      total_count:
        type: integer
      total_pages:
        type: integer
    type: object
  spinnet.CreateSpinnetRequest:
    properties:
      category:
        maxLength: 100
        type: string
      content:
        items:
          type: integer
        type: array
      name:
        maxLength: 255
        type: string
      owner_id:
        type: string
      smart_syntax:
        maxLength: 100
        type: string
    required:
    - name
    type: object
  spinnet.SpinnetListResponse:
    properties:
      meta:
        $ref: '#/definitions/queryparam.PaginationMeta'
      spinnets:
        items:
          $ref: '#/definitions/spinnet.SpinnetResponse'
        type: array
    type: object
  spinnet.SpinnetResponse:
    properties:
      category:
        type: string
      content:
        items:
          type: integer
        type: array
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      owner_id:
        type: string
      smart_syntax:
        type: string
      updated_at:
        type: string
    type: object
  spinnet.UpdateSpinnetRequest:
    properties:
      category:
        maxLength: 100
        type: string
      content:
        items:
          type: integer
        type: array
      name:
        maxLength: 255
        type: string
      smart_syntax:
        maxLength: 100
        type: string
    required:
    - name
    type: object
...
host: localhost:8000
info:
  contact:
    email: trao0312@gmail.com
    name: API Support
    url: https://github.com/justblue/samsa
  description: Samsa is a simple backend for a writing platform.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Samsa
  version: 0.1.0
paths:
...
  /spinnets:
    get:
      consumes:
      - application/json
      description: Retrieves a paginated list of spinnets.
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/spinnet.SpinnetListResponse'
      summary: List spinnets
      tags:
      - spinnets
    post:
      consumes:
      - application/json
      description: Creates a new spinnet template.
      parameters:
      - description: Spinnet creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/spinnet.CreateSpinnetRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/spinnet.SpinnetResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierror.APIError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/apierror.APIError'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/apierror.APIError'
      summary: Create spinnet
      tags:
      - spinnets
  /spinnets/{spinnet_id}:
    delete:
      description: Deletes a spinnet template.
      parameters:
      - description: Spinnet UUID
        in: path
        name: spinnet_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierror.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apierror.APIError'
      summary: Delete spinnet
      tags:
      - spinnets
    get:
      consumes:
      - application/json
      description: Retrieves a spinnet by its UUID.
      parameters:
      - description: Spinnet UUID
        in: path
        name: spinnet_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/spinnet.SpinnetResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierror.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apierror.APIError'
      summary: Get spinnet by ID
      tags:
      - spinnets
    put:
      consumes:
      - application/json
      description: Updates a spinnet template.
      parameters:
      - description: Spinnet UUID
        in: path
        name: spinnet_id
        required: true
        type: string
      - description: Spinnet update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/spinnet.UpdateSpinnetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/spinnet.SpinnetResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierror.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apierror.APIError'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/apierror.APIError'
      summary: Update spinnet
      tags:
      - spinnets
  /spinnets/syntax/{syntax}:
    get:
      consumes:
      - application/json
      description: Retrieves a spinnet by its smart syntax (e.g., /welcome).
      parameters:
      - description: Smart syntax (e.g., /welcome)
        in: path
        name: syntax
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/spinnet.SpinnetResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierror.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apierror.APIError'
      summary: Get spinnet by smart syntax
      tags:
      - spinnets
...
swagger: "2.0"